# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peruji API Documentation",
        "contact": {
            "email": "anjarpratama16@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Your Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "youremail@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Your Password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "youremail@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Your Password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contact-us": {
            "post": {
                "tags": [
                    "Landingpage"
                ],
                "operationId": "e4e9431d5df2f05ca258619dffbe0015",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "email",
                                    "phone",
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Your Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "youremail@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "628123123123"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Subject Message"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "This is my message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/event-registration": {
            "post": {
                "tags": [
                    "Landingpage"
                ],
                "operationId": "71c96003250b22f22d0bdd75e7a736b5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "eventregistrationname",
                                    "fullname",
                                    "email",
                                    "phone",
                                    "ofcphone",
                                    "company",
                                    "address"
                                ],
                                "properties": {
                                    "eventregistrationname": {
                                        "type": "string",
                                        "example": "Event Name"
                                    },
                                    "fullname": {
                                        "type": "string",
                                        "example": "Your Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "youremail@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "628123123123"
                                    },
                                    "ofcphone": {
                                        "type": "string",
                                        "example": "0211231231"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "My Company Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "This is My Address"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/membership": {
            "post": {
                "tags": [
                    "Landingpage"
                ],
                "operationId": "d6e2cae2326de75fe0a27b24e30d433c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "gender",
                                    "email",
                                    "phone",
                                    "dob",
                                    "org",
                                    "address",
                                    "website",
                                    "ofcemail"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Member Name"
                                    },
                                    "gender": {
                                        "description": "Gender: M = Male, F = Female",
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "membername@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "628123123123"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-30"
                                    },
                                    "org": {
                                        "type": "string",
                                        "example": "Member Organization"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "This is my address"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://mywebsite.com"
                                    },
                                    "ofcemail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "myoffice@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/news/{page}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "365eb74d54394e2fc10ef3cffc41af94",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page Number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "fd9358471b94ed4dc34d03ea4c06b645",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "newsname",
                                    "newsdate",
                                    "description",
                                    "photo",
                                    "additionalcontent"
                                ],
                                "properties": {
                                    "newsname": {
                                        "type": "string",
                                        "example": "News Title"
                                    },
                                    "newsdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-30"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "News Description"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "additionalcontent": {
                                        "type": "text",
                                        "example": "Some links or another"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-news/": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "3cf915edb95696298083208db11d1aef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "news",
                                    "newsname",
                                    "newsdate",
                                    "description",
                                    "additionalcontent"
                                ],
                                "properties": {
                                    "news": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "newsname": {
                                        "type": "string",
                                        "example": "News Title"
                                    },
                                    "newsdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-30"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "News Description"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "additionalcontent": {
                                        "type": "text",
                                        "example": "Some links or another"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Dashboard",
            "description": "Dashboard"
        },
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Landingpage",
            "description": "Landingpage"
        }
    ]
}